+++ src/scene/particlesystemcomponent.h
@@ -19,7 +19,7 @@
 
 struct ParticleSystemComponentData
 {
-  long flags;
+  int flags;
   lml::Bound3 bound;
   ParticleSystem *system;
   ParticleSystem::Instance const *instance;
@@ -38,7 +38,7 @@ class ParticleSystemComponentStorage : public BasicComponentStorage<ParticleSyst
 
   protected:
 
-    ParticleSystemComponentData *add(EntityId entity, lml::Bound3 const &bound, ParticleSystem *particlesystem, long flags);
+    ParticleSystemComponentData *add(EntityId entity, lml::Bound3 const &bound, ParticleSystem *particlesystem, int flags);
 
     void remove(EntityId entity) override;
 
@@ -69,14 +69,14 @@ class ParticleSystemComponent
     };
 
   public:
-    friend ParticleSystemComponent Scene::add_component<ParticleSystemComponent>(Scene::EntityId entity, ParticleSystem *particlesystem, long flags);
+    friend ParticleSystemComponent Scene::add_component<ParticleSystemComponent>(Scene::EntityId entity, ParticleSystem *particlesystem, int flags);
     friend ParticleSystemComponent Scene::get_component<ParticleSystemComponent>(Scene::EntityId entity);
 
   public:
     ParticleSystemComponent() = default;
     ParticleSystemComponent(ParticleSystemComponentData *data);
 
-    long flags() const { return m_data->flags; }
+    int flags() const { return m_data->flags; }
 
     lml::Bound3 const &bound() const { return m_data->bound; }
